Add test project, start building out test pages. Still need to finish clients, add content to test pages etc.

+++ src/AppleMusicAPI.NET/Models/Attributes/LibraryMusicVideoAttributes.cs
@@ -32,7 +32,7 @@ namespace AppleMusicAPI.NET.Models.Attributes
         /// <summary>
         /// The duration of the music video in milliseconds.
         /// </summary>
-        public long DurationInMillis { get; set; }
+        public int DurationInMillis { get; set; }
 
         /// <summary>
         /// (Required) The localized name of the music video.
